Move nested TimeZoneInfo types to their own files
