+++ src/mscorlib/src/System/TimeZoneInfo.cs
@@ -32,7 +32,7 @@ namespace System
     };
 
     [Serializable]
-    public sealed class TimeZoneInfo : IEquatable<TimeZoneInfo>, ISerializable, IDeserializationCallback
+    public sealed partial class TimeZoneInfo : IEquatable<TimeZoneInfo>, ISerializable, IDeserializationCallback
     {
         private enum TimeZoneInfoResult
         {
@@ -4294,1059 +4294,6 @@ namespace System
                         throw new InvalidTimeZoneException(Environment.GetResourceString("Argument_AdjustmentRulesOutOfOrder"));
                     }
                 }
-            }
-        }
-
-        [Serializable]
-        public sealed class AdjustmentRule : IEquatable<AdjustmentRule>, ISerializable, IDeserializationCallback
-        {
-            private readonly DateTime _dateStart;
-            private readonly DateTime _dateEnd;
-            private readonly TimeSpan _daylightDelta;
-            private readonly TransitionTime _daylightTransitionStart;
-            private readonly TransitionTime _daylightTransitionEnd;
-            private readonly TimeSpan _baseUtcOffsetDelta;   // delta from the default Utc offset (utcOffset = defaultUtcOffset + _baseUtcOffsetDelta)
-            private readonly bool _noDaylightTransitions;
-
-            public DateTime DateStart => _dateStart;
-
-            public DateTime DateEnd => _dateEnd;
-
-            public TimeSpan DaylightDelta => _daylightDelta;
-
-            public TransitionTime DaylightTransitionStart => _daylightTransitionStart;
-
-            public TransitionTime DaylightTransitionEnd => _daylightTransitionEnd;
-
-            internal TimeSpan BaseUtcOffsetDelta => _baseUtcOffsetDelta;
-
-            /// <summary>
-            /// Gets a value indicating that this AdjustmentRule fixes the time zone offset
-            /// from DateStart to DateEnd without any daylight transitions in between.
-            /// </summary>
-            internal bool NoDaylightTransitions => _noDaylightTransitions;
-
-            internal bool HasDaylightSaving =>
-                DaylightDelta != TimeSpan.Zero ||
-                (DaylightTransitionStart != default(TransitionTime) && DaylightTransitionStart.TimeOfDay != DateTime.MinValue) ||
-                (DaylightTransitionEnd != default(TransitionTime) && DaylightTransitionEnd.TimeOfDay != DateTime.MinValue.AddMilliseconds(1));
-
-            public bool Equals(AdjustmentRule other) =>
-                other != null &&
-                _dateStart == other._dateStart &&
-                _dateEnd == other._dateEnd &&
-                _daylightDelta == other._daylightDelta &&
-                _baseUtcOffsetDelta == other._baseUtcOffsetDelta &&
-                _daylightTransitionEnd.Equals(other._daylightTransitionEnd) &&
-                _daylightTransitionStart.Equals(other._daylightTransitionStart);
-
-            public override int GetHashCode() => _dateStart.GetHashCode();
-
-            private AdjustmentRule(
-                DateTime dateStart,
-                DateTime dateEnd,
-                TimeSpan daylightDelta,
-                TransitionTime daylightTransitionStart,
-                TransitionTime daylightTransitionEnd,
-                TimeSpan baseUtcOffsetDelta,
-                bool noDaylightTransitions)
-            {
-                ValidateAdjustmentRule(dateStart, dateEnd, daylightDelta,
-                       daylightTransitionStart, daylightTransitionEnd, noDaylightTransitions);
-
-                _dateStart = dateStart;
-                _dateEnd = dateEnd;
-                _daylightDelta = daylightDelta;
-                _daylightTransitionStart = daylightTransitionStart;
-                _daylightTransitionEnd = daylightTransitionEnd;
-                _baseUtcOffsetDelta = baseUtcOffsetDelta;
-                _noDaylightTransitions = noDaylightTransitions;
-            }
-
-            public static AdjustmentRule CreateAdjustmentRule(
-                DateTime dateStart,
-                DateTime dateEnd,
-                TimeSpan daylightDelta,
-                TransitionTime daylightTransitionStart,
-                TransitionTime daylightTransitionEnd)
-            {
-                return new AdjustmentRule(
-                    dateStart,
-                    dateEnd,
-                    daylightDelta,
-                    daylightTransitionStart,
-                    daylightTransitionEnd,
-                    baseUtcOffsetDelta: TimeSpan.Zero,
-                    noDaylightTransitions: false);
-            }
-
-            internal static AdjustmentRule CreateAdjustmentRule(
-                DateTime dateStart,
-                DateTime dateEnd,
-                TimeSpan daylightDelta,
-                TransitionTime daylightTransitionStart,
-                TransitionTime daylightTransitionEnd,
-                TimeSpan baseUtcOffsetDelta,
-                bool noDaylightTransitions)
-            {
-                return new AdjustmentRule(
-                    dateStart,
-                    dateEnd,
-                    daylightDelta,
-                    daylightTransitionStart,
-                    daylightTransitionEnd,
-                    baseUtcOffsetDelta,
-                    noDaylightTransitions);
-            }
-
-            //
-            // When Windows sets the daylight transition start Jan 1st at 12:00 AM, it means the year starts with the daylight saving on.
-            // We have to special case this value and not adjust it when checking if any date is in the daylight saving period.
-            //
-            internal bool IsStartDateMarkerForBeginningOfYear() =>
-                !NoDaylightTransitions &&
-                DaylightTransitionStart.Month == 1 && DaylightTransitionStart.Day == 1 && DaylightTransitionStart.TimeOfDay.Hour == 0 &&
-                DaylightTransitionStart.TimeOfDay.Minute == 0 && DaylightTransitionStart.TimeOfDay.Second == 0 &&
-                _dateStart.Year == _dateEnd.Year;
-
-            //
-            // When Windows sets the daylight transition end Jan 1st at 12:00 AM, it means the year ends with the daylight saving on.
-            // We have to special case this value and not adjust it when checking if any date is in the daylight saving period.
-            //
-            internal bool IsEndDateMarkerForEndOfYear() =>
-                !NoDaylightTransitions &&
-                DaylightTransitionEnd.Month == 1 && DaylightTransitionEnd.Day == 1 && DaylightTransitionEnd.TimeOfDay.Hour == 0 &&
-                DaylightTransitionEnd.TimeOfDay.Minute == 0 && DaylightTransitionEnd.TimeOfDay.Second == 0 &&
-                _dateStart.Year == _dateEnd.Year;
-
-            /// <summary>
-            /// Helper function that performs all of the validation checks for the actory methods and deserialization callback.
-            /// </summary>
-            private static void ValidateAdjustmentRule(
-                DateTime dateStart,
-                DateTime dateEnd,
-                TimeSpan daylightDelta,
-                TransitionTime daylightTransitionStart,
-                TransitionTime daylightTransitionEnd,
-                bool noDaylightTransitions)
-            {
-                if (dateStart.Kind != DateTimeKind.Unspecified && dateStart.Kind != DateTimeKind.Utc)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeKindMustBeUnspecifiedOrUtc"), nameof(dateStart));
-                }
-
-                if (dateEnd.Kind != DateTimeKind.Unspecified && dateEnd.Kind != DateTimeKind.Utc)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeKindMustBeUnspecifiedOrUtc"), nameof(dateEnd));
-                }
-
-                if (daylightTransitionStart.Equals(daylightTransitionEnd) && !noDaylightTransitions)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_TransitionTimesAreIdentical"), nameof(daylightTransitionEnd));
-                }
-
-                if (dateStart > dateEnd)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_OutOfOrderDateTimes"), nameof(dateStart));
-                }
-
-                // This cannot use UtcOffsetOutOfRange to account for the scenario where Samoa moved across the International Date Line,
-                // which caused their current BaseUtcOffset to be +13. But on the other side of the line it was UTC-11 (+1 for daylight).
-                // So when trying to describe DaylightDeltas for those times, the DaylightDelta needs
-                // to be -23 (what it takes to go from UTC+13 to UTC-10)
-                if (daylightDelta.TotalHours < -23.0 || daylightDelta.TotalHours > 14.0)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(daylightDelta), daylightDelta, Environment.GetResourceString("ArgumentOutOfRange_UtcOffset"));
-                }
-
-                if (daylightDelta.Ticks % TimeSpan.TicksPerMinute != 0)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_TimeSpanHasSeconds"), nameof(daylightDelta));
-                }
-
-                if (dateStart != DateTime.MinValue && dateStart.Kind == DateTimeKind.Unspecified && dateStart.TimeOfDay != TimeSpan.Zero)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeHasTimeOfDay"), nameof(dateStart));
-                }
-
-                if (dateEnd != DateTime.MaxValue && dateEnd.Kind == DateTimeKind.Unspecified && dateEnd.TimeOfDay != TimeSpan.Zero)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeHasTimeOfDay"), nameof(dateEnd));
-                }
-                Contract.EndContractBlock();
-            }
-
-            void IDeserializationCallback.OnDeserialization(object sender)
-            {
-                // OnDeserialization is called after each instance of this class is deserialized.
-                // This callback method performs AdjustmentRule validation after being deserialized.
-
-                try
-                {
-                    ValidateAdjustmentRule(_dateStart, _dateEnd, _daylightDelta,
-                                           _daylightTransitionStart, _daylightTransitionEnd, _noDaylightTransitions);
-                }
-                catch (ArgumentException e)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                }
-            }
-
-            void ISerializable.GetObjectData(SerializationInfo info, StreamingContext context)
-            {
-                if (info == null)
-                {
-                    throw new ArgumentNullException(nameof(info));
-                }
-                Contract.EndContractBlock();
-
-                info.AddValue("DateStart", _dateStart);
-                info.AddValue("DateEnd", _dateEnd);
-                info.AddValue("DaylightDelta", _daylightDelta);
-                info.AddValue("DaylightTransitionStart", _daylightTransitionStart);
-                info.AddValue("DaylightTransitionEnd", _daylightTransitionEnd);
-                info.AddValue("BaseUtcOffsetDelta", _baseUtcOffsetDelta);
-                info.AddValue("NoDaylightTransitions", _noDaylightTransitions);
-            }
-
-            AdjustmentRule(SerializationInfo info, StreamingContext context)
-            {
-                if (info == null)
-                {
-                    throw new ArgumentNullException(nameof(info));
-                }
-
-                _dateStart = (DateTime)info.GetValue("DateStart", typeof(DateTime));
-                _dateEnd = (DateTime)info.GetValue("DateEnd", typeof(DateTime));
-                _daylightDelta = (TimeSpan)info.GetValue("DaylightDelta", typeof(TimeSpan));
-                _daylightTransitionStart = (TransitionTime)info.GetValue("DaylightTransitionStart", typeof(TransitionTime));
-                _daylightTransitionEnd = (TransitionTime)info.GetValue("DaylightTransitionEnd", typeof(TransitionTime));
-
-                object o = info.GetValueNoThrow("BaseUtcOffsetDelta", typeof(TimeSpan));
-                if (o != null)
-                {
-                    _baseUtcOffsetDelta = (TimeSpan)o;
-                }
-
-                o = info.GetValueNoThrow("NoDaylightTransitions", typeof(bool));
-                if (o != null)
-                {
-                    _noDaylightTransitions = (bool)o;
-                }
-            }
-        }
-
-        [Serializable]
-        public struct TransitionTime : IEquatable<TransitionTime>, ISerializable, IDeserializationCallback
-        {
-            private readonly DateTime _timeOfDay;
-            private readonly byte _month;
-            private readonly byte _week;
-            private readonly byte _day;
-            private readonly DayOfWeek _dayOfWeek;
-            private readonly bool _isFixedDateRule;
-
-            public DateTime TimeOfDay => _timeOfDay;
-
-            public int Month => _month;
-
-            public int Week => _week;
-
-            public int Day => _day;
-
-            public DayOfWeek DayOfWeek => _dayOfWeek;
-
-            public bool IsFixedDateRule => _isFixedDateRule;
-
-            [Pure]
-            public override bool Equals(object obj) =>
-                obj is TransitionTime && Equals((TransitionTime)obj);
-
-            public static bool operator ==(TransitionTime t1, TransitionTime t2) => t1.Equals(t2);
-
-            public static bool operator !=(TransitionTime t1, TransitionTime t2) => !t1.Equals(t2);
-
-            [Pure]
-            public bool Equals(TransitionTime other) =>
-                _isFixedDateRule == other._isFixedDateRule &&
-                _timeOfDay == other._timeOfDay &&
-                _month == other._month &&
-                (other._isFixedDateRule ?
-                    _day == other._day :
-                    _week == other._week && _dayOfWeek == other._dayOfWeek);
-
-            public override int GetHashCode() => (int)_month ^ (int)_week << 8;
-
-            private TransitionTime(DateTime timeOfDay, int month, int week, int day, DayOfWeek dayOfWeek, bool isFixedDateRule)
-            {
-                ValidateTransitionTime(timeOfDay, month, week, day, dayOfWeek);
-
-                _timeOfDay = timeOfDay;
-                _month = (byte)month;
-                _week = (byte)week;
-                _day = (byte)day;
-                _dayOfWeek = dayOfWeek;
-                _isFixedDateRule = isFixedDateRule;
-            }
-
-            public static TransitionTime CreateFixedDateRule(DateTime timeOfDay, int month, int day) =>
-                new TransitionTime(timeOfDay, month, 1, day, DayOfWeek.Sunday, isFixedDateRule: true);
-
-            public static TransitionTime CreateFloatingDateRule(DateTime timeOfDay, int month, int week, DayOfWeek dayOfWeek) =>
-                new TransitionTime(timeOfDay, month, week, 1, dayOfWeek, isFixedDateRule: false);
-
-            /// <summary>
-            /// Helper function that validates a TransitionTime instance.
-            /// </summary>
-            private static void ValidateTransitionTime(DateTime timeOfDay, int month, int week, int day, DayOfWeek dayOfWeek)
-            {
-                if (timeOfDay.Kind != DateTimeKind.Unspecified)
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeKindMustBeUnspecified"), nameof(timeOfDay));
-                }
-
-                // Month range 1-12
-                if (month < 1 || month > 12)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(month), Environment.GetResourceString("ArgumentOutOfRange_MonthParam"));
-                }
-
-                // Day range 1-31
-                if (day < 1 || day > 31)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(day), Environment.GetResourceString("ArgumentOutOfRange_DayParam"));
-                }
-
-                // Week range 1-5
-                if (week < 1 || week > 5)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(week), Environment.GetResourceString("ArgumentOutOfRange_Week"));
-                }
-
-                // DayOfWeek range 0-6
-                if ((int)dayOfWeek < 0 || (int)dayOfWeek > 6)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(dayOfWeek), Environment.GetResourceString("ArgumentOutOfRange_DayOfWeek"));
-                }
-                Contract.EndContractBlock();
-
-                if (timeOfDay.Year != 1 || timeOfDay.Month != 1 || timeOfDay.Day != 1 || (timeOfDay.Ticks % TimeSpan.TicksPerMillisecond != 0))
-                {
-                    throw new ArgumentException(Environment.GetResourceString("Argument_DateTimeHasTicks"), nameof(timeOfDay));
-                }
-            }
-
-            void IDeserializationCallback.OnDeserialization(object sender)
-            {
-                // OnDeserialization is called after each instance of this class is deserialized.
-                // This callback method performs TransitionTime validation after being deserialized.
-
-                try
-                {
-                    ValidateTransitionTime(_timeOfDay, _month, _week, _day, _dayOfWeek);
-                }
-                catch (ArgumentException e)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                }
-            }
-
-            void ISerializable.GetObjectData(SerializationInfo info, StreamingContext context)
-            {
-                if (info == null)
-                {
-                    throw new ArgumentNullException(nameof(info));
-                }
-                Contract.EndContractBlock();
-
-                info.AddValue("TimeOfDay", _timeOfDay);
-                info.AddValue("Month", _month);
-                info.AddValue("Week", _week);
-                info.AddValue("Day", _day);
-                info.AddValue("DayOfWeek", _dayOfWeek);
-                info.AddValue("IsFixedDateRule", _isFixedDateRule);
-            }
-
-            TransitionTime(SerializationInfo info, StreamingContext context)
-            {
-                if (info == null)
-                {
-                    throw new ArgumentNullException(nameof(info));
-                }
-
-                _timeOfDay = (DateTime)info.GetValue("TimeOfDay", typeof(DateTime));
-                _month = (byte)info.GetValue("Month", typeof(byte));
-                _week = (byte)info.GetValue("Week", typeof(byte));
-                _day = (byte)info.GetValue("Day", typeof(byte));
-                _dayOfWeek = (DayOfWeek)info.GetValue("DayOfWeek", typeof(DayOfWeek));
-                _isFixedDateRule = (bool)info.GetValue("IsFixedDateRule", typeof(bool));
-            }
-        }
-
-        /// <summary>
-        /// This class is used to serialize and deserialize TimeZoneInfo
-        /// objects based on the custom string serialization format.
-        /// </summary>
-        private sealed class StringSerializer
-        {
-            private enum State
-            {
-                Escaped = 0,
-                NotEscaped = 1,
-                StartOfToken = 2,
-                EndOfLine = 3
-            }
-
-            private readonly string _serializedText;
-            private int _currentTokenStartIndex;
-            private State _state;
-
-            // the majority of the strings contained in the OS time zones fit in 64 chars
-            private const int InitialCapacityForString = 64;
-            private const char Esc = '\\';
-            private const char Sep = ';';
-            private const char Lhs = '[';
-            private const char Rhs = ']';
-            private const string EscString = "\\";
-            private const string SepString = ";";
-            private const string LhsString = "[";
-            private const string RhsString = "]";
-            private const string EscapedEsc = "\\\\";
-            private const string EscapedSep = "\\;";
-            private const string EscapedLhs = "\\[";
-            private const string EscapedRhs = "\\]";
-            private const string DateTimeFormat = "MM:dd:yyyy";
-            private const string TimeOfDayFormat = "HH:mm:ss.FFF";
-
-            /// <summary>
-            /// Creates the custom serialized string representation of a TimeZoneInfo instance.
-            /// </summary>
-            public static string GetSerializedString(TimeZoneInfo zone)
-            {
-                StringBuilder serializedText = StringBuilderCache.Acquire();
-
-                //
-                // <_id>;<_baseUtcOffset>;<_displayName>;<_standardDisplayName>;<_daylightDispayName>
-                //
-                serializedText.Append(SerializeSubstitute(zone.Id));
-                serializedText.Append(Sep);
-                serializedText.Append(SerializeSubstitute(
-                           zone.BaseUtcOffset.TotalMinutes.ToString(CultureInfo.InvariantCulture)));
-                serializedText.Append(Sep);
-                serializedText.Append(SerializeSubstitute(zone.DisplayName));
-                serializedText.Append(Sep);
-                serializedText.Append(SerializeSubstitute(zone.StandardName));
-                serializedText.Append(Sep);
-                serializedText.Append(SerializeSubstitute(zone.DaylightName));
-                serializedText.Append(Sep);
-
-                AdjustmentRule[] rules = zone.GetAdjustmentRules();
-
-                if (rules != null && rules.Length > 0)
-                {
-                    foreach (AdjustmentRule rule in rules)
-                    {
-                        serializedText.Append(Lhs);
-                        serializedText.Append(SerializeSubstitute(rule.DateStart.ToString(DateTimeFormat, DateTimeFormatInfo.InvariantInfo)));
-                        serializedText.Append(Sep);
-                        serializedText.Append(SerializeSubstitute(rule.DateEnd.ToString(DateTimeFormat, DateTimeFormatInfo.InvariantInfo)));
-                        serializedText.Append(Sep);
-                        serializedText.Append(SerializeSubstitute(rule.DaylightDelta.TotalMinutes.ToString(CultureInfo.InvariantCulture)));
-                        serializedText.Append(Sep);
-                        // serialize the TransitionTime's
-                        SerializeTransitionTime(rule.DaylightTransitionStart, serializedText);
-                        serializedText.Append(Sep);
-                        SerializeTransitionTime(rule.DaylightTransitionEnd, serializedText);
-                        serializedText.Append(Sep);
-                        if (rule.BaseUtcOffsetDelta != TimeSpan.Zero)
-                        {
-                            // Serialize it only when BaseUtcOffsetDelta has a value to reduce the impact of adding rule.BaseUtcOffsetDelta
-                            serializedText.Append(SerializeSubstitute(rule.BaseUtcOffsetDelta.TotalMinutes.ToString(CultureInfo.InvariantCulture)));
-                            serializedText.Append(Sep);
-                        }
-                        if (rule.NoDaylightTransitions)
-                        {
-                            // Serialize it only when NoDaylightTransitions is true to reduce the impact of adding rule.NoDaylightTransitions
-                            serializedText.Append(SerializeSubstitute("1"));
-                            serializedText.Append(Sep);
-                        }
-                        serializedText.Append(Rhs);
-                    }
-                }
-                serializedText.Append(Sep);
-                return StringBuilderCache.GetStringAndRelease(serializedText);
-            }
-
-            /// <summary>
-            /// Instantiates a TimeZoneInfo from a custom serialized string.
-            /// </summary>
-            public static TimeZoneInfo GetDeserializedTimeZoneInfo(string source)
-            {
-                StringSerializer s = new StringSerializer(source);
-
-                string id = s.GetNextStringValue(canEndWithoutSeparator: false);
-                TimeSpan baseUtcOffset = s.GetNextTimeSpanValue(canEndWithoutSeparator: false);
-                string displayName = s.GetNextStringValue(canEndWithoutSeparator: false);
-                string standardName = s.GetNextStringValue(canEndWithoutSeparator: false);
-                string daylightName = s.GetNextStringValue(canEndWithoutSeparator: false);
-                AdjustmentRule[] rules = s.GetNextAdjustmentRuleArrayValue(canEndWithoutSeparator: false);
-
-                try
-                {
-                    return new TimeZoneInfo(id, baseUtcOffset, displayName, standardName, daylightName, rules, disableDaylightSavingTime: false);
-                }
-                catch (ArgumentException ex)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), ex);
-                }
-                catch (InvalidTimeZoneException ex)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), ex);
-                }
-            }
-
-            private StringSerializer(string str)
-            {
-                _serializedText = str;
-                _state = State.StartOfToken;
-            }
-
-            /// <summary>
-            /// Returns a new string with all of the reserved sub-strings escaped
-            ///
-            /// ";" -> "\;"
-            /// "[" -> "\["
-            /// "]" -> "\]"
-            /// "\" -> "\\"
-            /// </summary>
-            private static string SerializeSubstitute(string text)
-            {
-                text = text.Replace(EscString, EscapedEsc);
-                text = text.Replace(LhsString, EscapedLhs);
-                text = text.Replace(RhsString, EscapedRhs);
-                return text.Replace(SepString, EscapedSep);
-            }
-
-            /// <summary>
-            /// Helper method to serialize a TimeZoneInfo.TransitionTime object.
-            /// </summary>
-            private static void SerializeTransitionTime(TransitionTime time, StringBuilder serializedText)
-            {
-                serializedText.Append(Lhs);
-                int fixedDate = (time.IsFixedDateRule ? 1 : 0);
-                serializedText.Append(fixedDate.ToString(CultureInfo.InvariantCulture));
-                serializedText.Append(Sep);
-
-                if (time.IsFixedDateRule)
-                {
-                    serializedText.Append(SerializeSubstitute(time.TimeOfDay.ToString(TimeOfDayFormat, DateTimeFormatInfo.InvariantInfo)));
-                    serializedText.Append(Sep);
-                    serializedText.Append(SerializeSubstitute(time.Month.ToString(CultureInfo.InvariantCulture)));
-                    serializedText.Append(Sep);
-                    serializedText.Append(SerializeSubstitute(time.Day.ToString(CultureInfo.InvariantCulture)));
-                    serializedText.Append(Sep);
-                }
-                else
-                {
-                    serializedText.Append(SerializeSubstitute(time.TimeOfDay.ToString(TimeOfDayFormat, DateTimeFormatInfo.InvariantInfo)));
-                    serializedText.Append(Sep);
-                    serializedText.Append(SerializeSubstitute(time.Month.ToString(CultureInfo.InvariantCulture)));
-                    serializedText.Append(Sep);
-                    serializedText.Append(SerializeSubstitute(time.Week.ToString(CultureInfo.InvariantCulture)));
-                    serializedText.Append(Sep);
-                    serializedText.Append(SerializeSubstitute(((int)time.DayOfWeek).ToString(CultureInfo.InvariantCulture)));
-                    serializedText.Append(Sep);
-                }
-                serializedText.Append(Rhs);
-            }
-
-            /// <summary>
-            /// Helper function to determine if the passed in string token is allowed to be preceeded by an escape sequence token.
-            /// </summary>
-            private static void VerifyIsEscapableCharacter(char c)
-            {
-                if (c != Esc && c != Sep && c != Lhs && c != Rhs)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidEscapeSequence", c));
-                }
-            }
-
-            /// <summary>
-            /// Helper function that reads past "v.Next" data fields. Receives a "depth" parameter indicating the
-            /// current relative nested bracket depth that _currentTokenStartIndex is at. The function ends
-            /// successfully when "depth" returns to zero (0).
-            /// </summary>
-            private void SkipVersionNextDataFields(int depth /* starting depth in the nested brackets ('[', ']')*/)
-            {
-                if (_currentTokenStartIndex < 0 || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                State tokenState = State.NotEscaped;
-
-                // walk the serialized text, building up the token as we go...
-                for (int i = _currentTokenStartIndex; i < _serializedText.Length; i++)
-                {
-                    if (tokenState == State.Escaped)
-                    {
-                        VerifyIsEscapableCharacter(_serializedText[i]);
-                        tokenState = State.NotEscaped;
-                    }
-                    else if (tokenState == State.NotEscaped)
-                    {
-                        switch (_serializedText[i])
-                        {
-                            case Esc:
-                                tokenState = State.Escaped;
-                                break;
-
-                            case Lhs:
-                                depth++;
-                                break;
-                            case Rhs:
-                                depth--;
-                                if (depth == 0)
-                                {
-                                    _currentTokenStartIndex = i + 1;
-                                    if (_currentTokenStartIndex >= _serializedText.Length)
-                                    {
-                                        _state = State.EndOfLine;
-                                    }
-                                    else
-                                    {
-                                        _state = State.StartOfToken;
-                                    }
-                                    return;
-                                }
-                                break;
-
-                            case '\0':
-                                // invalid character
-                                throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-
-                            default:
-                                break;
-                        }
-                    }
-                }
-
-                throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-            }
-
-            /// <summary>
-            /// Helper function that reads a string token from the serialized text. The function
-            /// updates <see cref="_currentTokenStartIndex"/> to point to the next token on exit.
-            /// Also <see cref="_state"/> is set to either <see cref="State.StartOfToken"/> or
-            /// <see cref="State.EndOfLine"/> on exit.
-            /// </summary>
-            /// <param name="canEndWithoutSeparator">
-            /// - When set to 'false' the function requires the string token end with a ";".
-            /// - When set to 'true' the function requires that the string token end with either
-            ///   ";", <see cref="State.EndOfLine"/>, or "]". In the case that "]" is the terminal
-            ///   case the <see cref="_currentTokenStartIndex"/> is left pointing at index "]" to
-            ///   allow the caller to update its depth logic.
-            /// </param>
-            private string GetNextStringValue(bool canEndWithoutSeparator)
-            {
-                // first verify the internal state of the object
-                if (_state == State.EndOfLine)
-                {
-                    if (canEndWithoutSeparator)
-                    {
-                        return null;
-                    }
-                    else
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                    }
-                }
-                if (_currentTokenStartIndex < 0 || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                State tokenState = State.NotEscaped;
-                StringBuilder token = StringBuilderCache.Acquire(InitialCapacityForString);
-
-                // walk the serialized text, building up the token as we go...
-                for (int i = _currentTokenStartIndex; i < _serializedText.Length; i++)
-                {
-                    if (tokenState == State.Escaped)
-                    {
-                        VerifyIsEscapableCharacter(_serializedText[i]);
-                        token.Append(_serializedText[i]);
-                        tokenState = State.NotEscaped;
-                    }
-                    else if (tokenState == State.NotEscaped)
-                    {
-                        switch (_serializedText[i])
-                        {
-                            case Esc:
-                                tokenState = State.Escaped;
-                                break;
-
-                            case Lhs:
-                                // '[' is an unexpected character
-                                throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-
-                            case Rhs:
-                                if (canEndWithoutSeparator)
-                                {
-                                    // if ';' is not a required terminal then treat ']' as a terminal
-                                    // leave _currentTokenStartIndex pointing to ']' so our callers can handle
-                                    // this special case
-                                    _currentTokenStartIndex = i;
-                                    _state = State.StartOfToken;
-                                    return token.ToString();
-                                }
-                                else
-                                {
-                                    // ']' is an unexpected character
-                                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                                }
-
-                            case Sep:
-                                _currentTokenStartIndex = i + 1;
-                                if (_currentTokenStartIndex >= _serializedText.Length)
-                                {
-                                    _state = State.EndOfLine;
-                                }
-                                else
-                                {
-                                    _state = State.StartOfToken;
-                                }
-                                return StringBuilderCache.GetStringAndRelease(token);
-
-                            case '\0':
-                                // invalid character
-                                throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-
-                            default:
-                                token.Append(_serializedText[i]);
-                                break;
-                        }
-                    }
-                }
-                //
-                // we are at the end of the line
-                //
-                if (tokenState == State.Escaped)
-                {
-                    // we are at the end of the serialized text but we are in an escaped state
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidEscapeSequence", string.Empty));
-                }
-
-                if (!canEndWithoutSeparator)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                _currentTokenStartIndex = _serializedText.Length;
-                _state = State.EndOfLine;
-                return StringBuilderCache.GetStringAndRelease(token);
-            }
-
-            /// <summary>
-            /// Helper function to read a DateTime token.
-            /// </summary>
-            private DateTime GetNextDateTimeValue(bool canEndWithoutSeparator, string format)
-            {
-                string token = GetNextStringValue(canEndWithoutSeparator);
-                DateTime time;
-                if (!DateTime.TryParseExact(token, format, DateTimeFormatInfo.InvariantInfo, DateTimeStyles.None, out time))
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                return time;
-            }
-
-            /// <summary>
-            /// Helper function to read a TimeSpan token.
-            /// </summary>
-            private TimeSpan GetNextTimeSpanValue(bool canEndWithoutSeparator)
-            {
-                int token = GetNextInt32Value(canEndWithoutSeparator);
-                try
-                {
-                    return new TimeSpan(hours: 0, minutes: token, seconds: 0);
-                }
-                catch (ArgumentOutOfRangeException e)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                }
-            }
-
-            /// <summary>
-            /// Helper function to read an Int32 token.
-            /// </summary>
-            private int GetNextInt32Value(bool canEndWithoutSeparator)
-            {
-                string token = GetNextStringValue(canEndWithoutSeparator);
-                int value;
-                if (!int.TryParse(token, NumberStyles.AllowLeadingSign /* "[sign]digits" */, CultureInfo.InvariantCulture, out value))
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                return value;
-            }
-
-            /// <summary>
-            /// Helper function to read an AdjustmentRule[] token.
-            /// </summary>
-            private AdjustmentRule[] GetNextAdjustmentRuleArrayValue(bool canEndWithoutSeparator)
-            {
-                List<AdjustmentRule> rules = new List<AdjustmentRule>(1);
-                int count = 0;
-
-                // individual AdjustmentRule array elements do not require semicolons
-                AdjustmentRule rule = GetNextAdjustmentRuleValue(canEndWithoutSeparator: true);
-                while (rule != null)
-                {
-                    rules.Add(rule);
-                    count++;
-
-                    rule = GetNextAdjustmentRuleValue(canEndWithoutSeparator: true);
-                }
-
-                if (!canEndWithoutSeparator)
-                {
-                    // the AdjustmentRule array must end with a separator
-                    if (_state == State.EndOfLine)
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                    }
-                    if (_currentTokenStartIndex < 0 || _currentTokenStartIndex >= _serializedText.Length)
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                    }
-                }
-
-                return count != 0 ? rules.ToArray() : null;
-            }
-
-            /// <summary>
-            /// Helper function to read an AdjustmentRule token.
-            /// </summary>
-            private AdjustmentRule GetNextAdjustmentRuleValue(bool canEndWithoutSeparator)
-            {
-                // first verify the internal state of the object
-                if (_state == State.EndOfLine)
-                {
-                    if (canEndWithoutSeparator)
-                    {
-                        return null;
-                    }
-                    else
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                    }
-                }
-
-                if (_currentTokenStartIndex < 0 || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                // check to see if the very first token we see is the separator
-                if (_serializedText[_currentTokenStartIndex] == Sep)
-                {
-                    return null;
-                }
-
-                // verify the current token is a left-hand-side marker ("[")
-                if (_serializedText[_currentTokenStartIndex] != Lhs)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                _currentTokenStartIndex++;
-
-                DateTime dateStart = GetNextDateTimeValue(false, DateTimeFormat);
-                DateTime dateEnd = GetNextDateTimeValue(false, DateTimeFormat);
-                TimeSpan daylightDelta = GetNextTimeSpanValue(canEndWithoutSeparator: false);
-                TransitionTime daylightStart = GetNextTransitionTimeValue(canEndWithoutSeparator: false);
-                TransitionTime daylightEnd = GetNextTransitionTimeValue(canEndWithoutSeparator: false);
-                TimeSpan baseUtcOffsetDelta = TimeSpan.Zero;
-                int noDaylightTransitions = 0;
-
-                // verify that the string is now at the right-hand-side marker ("]") ...
-
-                if (_state == State.EndOfLine || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                // Check if we have baseUtcOffsetDelta in the serialized string and then deserialize it
-                if ((_serializedText[_currentTokenStartIndex] >= '0' && _serializedText[_currentTokenStartIndex] <= '9') ||
-                    _serializedText[_currentTokenStartIndex] == '-' || _serializedText[_currentTokenStartIndex] == '+')
-                {
-                    baseUtcOffsetDelta = GetNextTimeSpanValue(canEndWithoutSeparator: false);
-                }
-
-                // Check if we have NoDaylightTransitions in the serialized string and then deserialize it
-                if ((_serializedText[_currentTokenStartIndex] >= '0' && _serializedText[_currentTokenStartIndex] <= '1'))
-                {
-                    noDaylightTransitions = GetNextInt32Value(canEndWithoutSeparator: false);
-                }
-
-                if (_state == State.EndOfLine || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                if (_serializedText[_currentTokenStartIndex] != Rhs)
-                {
-                    // skip ahead of any "v.Next" data at the end of the AdjustmentRule
-                    //
-                    // FUTURE: if the serialization format is extended in the future then this
-                    // code section will need to be changed to read the new fields rather
-                    // than just skipping the data at the end of the [AdjustmentRule].
-                    SkipVersionNextDataFields(1);
-                }
-                else
-                {
-                    _currentTokenStartIndex++;
-                }
-
-                // create the AdjustmentRule from the deserialized fields ...
-
-                AdjustmentRule rule;
-                try
-                {
-                    rule = AdjustmentRule.CreateAdjustmentRule(dateStart, dateEnd, daylightDelta, daylightStart, daylightEnd, baseUtcOffsetDelta, noDaylightTransitions > 0);
-                }
-                catch (ArgumentException e)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                }
-
-                // finally set the state to either EndOfLine or StartOfToken for the next caller
-                if (_currentTokenStartIndex >= _serializedText.Length)
-                {
-                    _state = State.EndOfLine;
-                }
-                else
-                {
-                    _state = State.StartOfToken;
-                }
-                return rule;
-            }
-
-            /// <summary>
-            /// Helper function to read a TransitionTime token.
-            /// </summary>
-            private TransitionTime GetNextTransitionTimeValue(bool canEndWithoutSeparator)
-            {
-                // first verify the internal state of the object
-
-                if (_state == State.EndOfLine ||
-                    (_currentTokenStartIndex < _serializedText.Length && _serializedText[_currentTokenStartIndex] == Rhs))
-                {
-                    //
-                    // we are at the end of the line or we are starting at a "]" character
-                    //
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                if (_currentTokenStartIndex < 0 || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                // verify the current token is a left-hand-side marker ("[")
-
-                if (_serializedText[_currentTokenStartIndex] != Lhs)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-                _currentTokenStartIndex++;
-
-                int isFixedDate = GetNextInt32Value(canEndWithoutSeparator: false);
-
-                if (isFixedDate != 0 && isFixedDate != 1)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                TransitionTime transition;
-
-                DateTime timeOfDay = GetNextDateTimeValue(false, TimeOfDayFormat);
-                timeOfDay = new DateTime(1, 1, 1, timeOfDay.Hour, timeOfDay.Minute, timeOfDay.Second, timeOfDay.Millisecond);
-
-                int month = GetNextInt32Value(canEndWithoutSeparator: false);
-
-                if (isFixedDate == 1)
-                {
-                    int day = GetNextInt32Value(canEndWithoutSeparator: false);
-
-                    try
-                    {
-                        transition = TransitionTime.CreateFixedDateRule(timeOfDay, month, day);
-                    }
-                    catch (ArgumentException e)
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                    }
-                }
-                else
-                {
-                    int week = GetNextInt32Value(canEndWithoutSeparator: false);
-                    int dayOfWeek = GetNextInt32Value(canEndWithoutSeparator: false);
-
-                    try
-                    {
-                        transition = TransitionTime.CreateFloatingDateRule(timeOfDay, month, week, (DayOfWeek)dayOfWeek);
-                    }
-                    catch (ArgumentException e)
-                    {
-                        throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"), e);
-                    }
-                }
-
-                // verify that the string is now at the right-hand-side marker ("]") ...
-
-                if (_state == State.EndOfLine || _currentTokenStartIndex >= _serializedText.Length)
-                {
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                if (_serializedText[_currentTokenStartIndex] != Rhs)
-                {
-                    // skip ahead of any "v.Next" data at the end of the AdjustmentRule
-                    //
-                    // FUTURE: if the serialization format is extended in the future then this
-                    // code section will need to be changed to read the new fields rather
-                    // than just skipping the data at the end of the [TransitionTime].
-                    SkipVersionNextDataFields(1);
-                }
-                else
-                {
-                    _currentTokenStartIndex++;
-                }
-
-                // check to see if the string is now at the separator (";") ...
-                bool sepFound = false;
-                if (_currentTokenStartIndex < _serializedText.Length &&
-                    _serializedText[_currentTokenStartIndex] == Sep)
-                {
-                    // handle the case where we ended on a ";"
-                    _currentTokenStartIndex++;
-                    sepFound = true;
-                }
-
-                if (!sepFound && !canEndWithoutSeparator)
-                {
-                    // we MUST end on a separator
-                    throw new SerializationException(Environment.GetResourceString("Serialization_InvalidData"));
-                }
-
-                // finally set the state to either EndOfLine or StartOfToken for the next caller
-                if (_currentTokenStartIndex >= _serializedText.Length)
-                {
-                    _state = State.EndOfLine;
-                }
-                else
-                {
-                    _state = State.StartOfToken;
-                }
-                return transition;
             }
         }
